+++ src/rpc/mod.rs
@@ -173,6 +173,8 @@ macro_rules! for_each_method {
         $callback!(crate::rpc::state::StateGetClaims);
         $callback!(crate::rpc::state::StateGetAllocation);
         $callback!(crate::rpc::state::StateGetAllocations);
+        $callback!(crate::rpc::state::StateGetAllocationIdForPendingDeal);
+        $callback!(crate::rpc::state::StateGetAllocationForPendingDeal);
         $callback!(crate::rpc::state::StateSectorExpiration);
         $callback!(crate::rpc::state::StateSectorPartition);
         $callback!(crate::rpc::state::StateLookupRobustAddress);
@@ -333,6 +335,10 @@ pub struct RPCState<DB> {
 impl<DB: Blockstore> RPCState<DB> {
     pub fn store(&self) -> &DB {
         self.chain_store.blockstore()
+    }
+
+    pub fn store_owned(&self) -> Arc<DB> {
+        self.state_manager.blockstore_owned()
     }
 }